Clear circular dependencies
